Remove CodeStubArguments::SetArgumentValue
The API is somewhat dangerous for the sandbox since it writes directly
into the stack. As such, if used incorrectly (e.g. when there is
confusion about the number of available arguments), it may cause stack
memory corruption. Since the API doesn't seem to be in use at the
moment, this CL simply removes it.

Bug: 412741811

Change-Id: Ibf972f3aac04f0ec90cddbe076d1f265aa4e028d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6485689
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99914}