[inspector] Surface 'debugId' magic comment as 'buildId' CDP field
We pass the field straight-through from V8 via CDP.

The only quirk is when scripts fail to parse: In this case we have some
existing logic that uses a best-effort regex to retrieve the magic
comments. To stay consistent, we'll do the same for `debugId`.

R=leszeks@chromium.org, pfaffe@chromium.org

Bug: 413177407
Change-Id: Ic1793470f3e98b225a1ff907ca613a4953eb5db2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6491466
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99907}