[sandbox] Fix OOB stack read in ArrayPrototypePush
ArrayPrototypePush will first try a fast path that pushes arguments into
the array as long as their type is compatible with the array's elements
kind. If an incompatible value is encountered, the fast path bails out
to a slow path that transitions the array. For that, it is necessary to
determine how many arguments have already been pushed, so that the slow
path only processes the remaining arguments. The logic for that used to
work by re-loading the length of the JSArray and assuming that the
difference of the new length and the original length was the number of
processed arguments. However, the length can change independently in
between the two reads, leading to a subsequent out-of-bounds read from
the stack when accessing the remaining arguments. While this isn't
technically a sandbox violation (no out-of-sandbox memory is corrupted),
it's still worth fixing these issues for fuzzer cleanliness and to
prevent these bugs from becoming real security issues in the future due
to code refactoring or similar. The fix implemented in this CL simply
avoids the double-load of the .length property from the array.

Bug: 412741811
Change-Id: I432f22a45e3be21bf210c3008dc1e0d94573df0b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6490467
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99911}