[code] SBX_CHECK code kind in some getters
The deopt data or bytecode field has a value dependent on the kind of
the code; since the Code object is trusted, the value in this field can
be trusted to be consistent with the Code kind, however the reference to
the Code object itself might be from an untrusted object (via the EPT),
and there we might get a confusion between code kinds.

To avoid this confusion causing issues, SBX_CHECK the code kind in the
getter of this field.

Bug: 412587816
Change-Id: Ibc92885b898b573200099ce70cf5e3b5093a032c
Fixed: 413427118
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6491089
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99910}