[wasm-custom-desc] New WasmExportWrapperCache
This replaces the cache we use for JS-to-Wasm wrappers, which so far
was a WeakFixedArray indexed by canonical type index, with a weak
hash map: still based on a WeakFixedArray, but indexing like a
HashTable (open addressing, quadratic probing). The intention is
twofold:
(1) Save memory. We expect that there may be tens of thousands of
    canonical types, but only a small minority of them will get
    compiled js-to-wasm wrappers.
(2) Support more than one wrapper per type index for an upcoming new
    feature: exported Wasm functions will gain a configurable ability
    to pass the JS-side receiver as the first Wasm-side parameter,
    rather than ignoring it as they did so far. That behavior will
    need to be baked into the compiled wrappers.

Bug: 403372470
Change-Id: I524761b4e1cfdd32113812d03488b446c9a4ba88
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6491348
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99941}