[wasm-custom-desc] Support "receiver is first param"
This adds a new capability to exported Wasm functions: when they
set the "receiver is first param" bit, then the JS-side receiver
of a call to them will be passed on to Wasm as the first parameter.
The JS-side parameters will accordingly be shifted up by one. For
example, `foo.bar(baz)` will call a Wasm function with parameters
(foo, baz).
This new feature is not exposed yet in this CL, that will follow.

Bug: 403372470
Change-Id: I36a2bcc2f98886103a225907246255947bf78547
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6491528
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99949}