[revec][wasm] Fix extract node from early emitted input tree
When force-packing nodes, it's possible to reduce the inputs of the
other node in advance. Some input can be part of a PackNode and create
additional Extract128 node. We will skip such PackNode if it's already
reduced in later visit. This patch adds a check to avoid generating
redundant Extract128 node in such case.

Also fixed the asan failure on imms buffer overflow with the new test.

Bug: chromium:412732875
Change-Id: Ica86e8332db43132ff42480994019186b4137fca
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6492222
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Yolanda Chen <yolanda.chen@intel.com>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100106}