[debug] Use kNoContext instead of kInvalidContext
Debug code in the c-entry stub set the context field on the isolate to
kInvalidContext when returning to generated code. This debug code should
help with finding C++ code that accesses a stale context, i.e. C++ code
accesses the context, but the generated code calling the C++ code forgot
to set the current context.

The problem with the debug code was that it set the context field to a
value that was not a valid tagged value, and when the GC would see that
value, it would crash.

With this CL, kNoContext is used instead of kInvalidContext. Thereby C++
code that access the context will also fail, but if the GC sees the
value, it can handle it.

R=leszeks@chromium.org

Fixed: 409593622
Change-Id: I21ce18de1afdda290df44319f621c7122da93794
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6494497
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99995}