[serializer] Promote ObjectTemplateInfo to RO space when possible
This CL adds v8::ObjectTemplate::SealAndPrepareForPromotionToReadOnly()
which is supposed to be called by embedder when the object template
is ready and is not going to be modified.

Additionally, support 3-way promo decisions for promotion to read only
space. The third mode "match host" is supposed to indicate that
decision depends on the outer object. This is useful for types such as
ArrayList or JSExternalObject (AKA v8::External), when it's impossible
to make a promo decision by looking just at object without knowing how
it's used.

Bug: 399393891
Fixed: 413385733
Change-Id: Ic99783ea69d293c5659c3a5d74cd7151d93ba54a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6416966
Auto-Submit: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99991}