[sandbox] Add hardware sandbox initialization test
As a follow-up to https://crrev.com/c/6494565, this new test now makes
sure that hardware support is correctly enabled if the system supports
memory protection keys. This requires changing the
MemoryProtectionKey::HasMemoryProtectionKeySupport routine to not only
check if the pkey glibc routines are available but to also check if
pkeys can actually be allocated (meaning both the CPU and the kernel
also support PKEYs). To better reflect the fact that this routine checks
whether the current system supports pkeys, this CL also renames it to
MemoryProtectionKey::SystemSupportsMemoryProtectionKeys.

Bug: 350324877
Change-Id: If70cd5b496d0e0d188c2885ab0a01f301b9f00c1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6500738
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#100024}