[wasm][stacks] Add guard pages around stacks
To ensure that a stack overflow or underflow cannot be exploited.

We are about to reduce the default initial stack size and stack limit
margin for growable stacks, which will make it more likely that we
accidentally hit these limits.

In the future, if these guard pages add too much memory overhead in
practice and we gained enough confidence in the stability of growable
stacks, we can consider removing them again.

R=clemensb@chromium.org

Bug: 42204615
Change-Id: Ia5e6b0f357857eef0f01cdbc4615f70b9438d18b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6507705
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100054}