[parser] Set start/end on materialized scopes
debug-evaluate requires the 'UniqueIdInScript' of scopes to compare
re-parsed with runtime scopes. parsing::ParseFunction materializes
outer scopes from runtime scopes before (re-)parsing a function.
Currently, this materialization step does not re-materialize the
start/end position, so debug-evaluate is not able to properly compare
parsed with runtime scopes.

This CL fixes this by also re-materializing the position info when
parsing functions.

For the script scope we don't have to set the start/end
position as the script scope always has a UniqueIdInScript of -2.

Note: The current implementation leads `HasContext()` to return
`false` even though we have a context and the parsed and runtime
scope. One could think that we might show variables as unavailable
in this case. But we got lucky: For the scope view, we re-parse the
full script, which results in full position information for all
scopes. We use `parsing::ParseFunction` only for debug-evaluate
where `HasContext()` for outer scopes is unused and does not matter.

R=leszeks@chromium.org

Bug: 399002824
Change-Id: If8c4e73693b112dbced945f2094730cd92b535c3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6507407
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100041}