[inspector] Extend 'step-over' test coverage
This CL takes two existing tests (step-into.js and for-of-loops.js) and
also steps through the examples with 'step-over'. Instead of repeatedly
doing a 'step into' until we are done, we now do one 'step into' to
get into the test function and then repeatedly 'step-over' until we
are done.

We want to change for-of loop stepping and having some additional test
coverage won't hurt.

R=bmeurer@chromium.org

Bug: 407773675
Change-Id: Iad393830742e4efb9b04ea1cb855a647ecf2ec2b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6509410
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100045}