[wasm] Take NativeModule lock before allocating fast api signature
For fast API calls from WebAssembly, a copy of the API signature is
created during module instantiation. This copy is stored in the
`signature_zone` of the WasmModule. As multiple workers can instantiate
the same wasm module concurrently, the `signature_zone` can be used
concurrently to allocation memory for signature copies, and therefore
the `signature_zone` has to be protected by a lock.

This CL introduces a scope such that the allocation lock of the
NativeModule can be taken outside the NativeModule implementation. An
alternative would be to create the copy of the signature within the
NativeModule implementation. This was not done because it would create a
dependency of the NativeModule implementation from the fast API.

R=clemensb@chromium.org

Bug: 397459607
Change-Id: Id3f688c7d6db5517f3ff13046dfcf835c78b65ba
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6508725
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100047}