[runtime] Fix flushing bug in InstallBaselineCode()
CreateAndAttachFeedbackVector() might trigger a GC which might in turn
flush the baseline generated code before it is installed.

This CL introduces an IsBaselineCompiledScope which keeps alive the
baseline code if there is one and prevents flushing. This scope is
then used in InstallBaselineCode() around
CreateAndAttachFeedbackVector().

Bug: 40270831, 413660243
Change-Id: Ic2dd6da4dbcd51bfd9e8f3afaec4f3ffccdac73c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6508505
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100049}