[wasm][growable-stacks] Reduce initial stack size
Reduce the default initial stack size for growable stacks.

Now that we always switch to the central stack for runtime calls, we
can also reduce the JS stack limit margin. It only needs to be large
enough to run C calls and builtins.

Also fix one issue that this change uncovered: when we allocate a new
segment, we need to ensure that the new segment is large enough to
contain the frame that is being evicted.

R=clemensb@chromium.org

Bug: 42204615
Change-Id: Ifdd262610a6cb13dc7a1b88cf130e2e8fb8e3226
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6508928
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100059}