[debug] Change for-of/for-in/for-await stepping behavior
This CL changes the stepping behavior for the following 'for' loops:

  for (const each of subject) {
  for (const each in subject) {
  for await (const each of subject) {

Currently we stop before the iterator 'next()' call (or before the
'done' check in the 'in' case). This leads to unintuitive stepping
behavior: For a loop with N iterations, we'll pause N + 1 times on
the 'const each' declaration.

This CL changes the iterator 'next()' call to a unbreakable source
position (to preserve stack traces) and makes the assignment to
the 'each' variable a breakable position instead.

This way, if a loop iterates N times, we'll pause N times on the
'const each' declaration.

R=bmeurer@chromium.org, leszeks@chromium.org

Fixed: 407773675
Change-Id: Id013447a1edeb39bbbffffab71922405e220afed
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6508377
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100058}