Load JS dispatch table from isolate.
The wasm wrappers are stored in a process-wide cache, but they use an
external reference to reference the js dispatch table, which is not
process-wide, but rather per-isolate-group/sandbox.

With this change the js dispatch table is normally loaded via a register
from the isolate data. That way the generated code is sandbox
independent and the wrappers can be stored in a process-wide cache.

Bug: 413385716
Bug: 42204201
Change-Id: I6c9ebd09eaa8ad88befcd7f736da3be1aa002c4c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6494801
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100070}