[wasm][deoptimizer] Resolve feedback vector in WasmLiftoffDeoptFinish only
Previously the deoptimizer would try to lookup the feedback vector in
the vector of feedback vectors in the wasm trusted instance.
This was implemented at a time when I still assumed that there will
always be a valid feedback vector, however due to multi-instance
scenarios, this turned out to be incorrect. Later on this was fixed in
the WasmLiftoffDeoptFinish to allocate and fill in those feedback
vectors for which we didn't find one yet.

Moving the lookup of the feedback vector to the builtin exclusively
simplifies the code and also removes an access to the heap from within
the deoptimizer (which we'd ideally reduce down to zero if possible).

Cq-Include-Trybots: luci.v8.try:v8_linux64_pku_dbg,v8_linux64_pku_rel
Bug: 415707239, 42204618
Change-Id: Iabe27b277bc49acfac6763bc135b301a36967255
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6511755
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100077}