Add separate builders for PKU testers
This adds (compiling) builders for the two PKU configurations.
Previously, the default release and debug builders were shared.
Since the src-side configs for GN are now newly checked in for these
new builders, we need to set the supported milestone to the current.
As an effect, the testers currently on beta/stable are removed until
the the next two milestones.

Bug: 415298009
Change-Id: I91c5b284526073a501ec9832dc75ef197e3d9e3a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6513780
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>