[osr] Only deopt osr code if deopt exits are inside the loop
DeoptAllOsrLoopsContainingDeoptExit was written with the intention of
capturing all osr codes which can reach the deopt exit. However, (a)
maglev osr never jumps up over the osr entry and (b) even turbofan osr
code might still be useful to clear a loop as long as the deopt is not
inside the loop itself.

Also, ensure that when we clear osr code we correctly reset the flags
on the feedback vector.

Bug: 416431775

Change-Id: Ife433a2ef233291d534db9bf6ae321ba19727321
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6516174
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100152}