[json] Parser: Add more checks for fast-iterable keys
We didn't check all the required properties for fast iterable keys in
the JSON parser. Some of these properties are only relevenat for
JSON.stringify, but nevertheless need to be checked in the parser before
we mark a descriptor array as fast iterable.

Fixed: 414962716,415551707
Bug: 411351808
Change-Id: Ib5900aec93e50e1412a3c9c1b17f792c679fdd24
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6511763
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100115}