[heap] Add optional "may grow" hint for heap allocations
This CL adds the capability to pass optional hints to heap allocations.
In particular this CL adds a "may grow" hint but we can use this for
other features in the future.

The motivation behind "may grow" is the resizing of large objects
that we now support using TryResizeLargeObject(). With out current
"first fit" allocation strategy in the region allocation, we simply
allocate a large object in the first free region large enough for the
allocation. However, this might not work well with resizing because
usually we start with a small allocation which we grow later if
necessary. We can now run into situations were resizing often fails
because we allocated in a small region initially with no room left
for growing.

The "may grow" hint should solve this by instead allocating in the
largest free region instead which should enable us to grow the
object to its final size successfully. Since we only grow the
last allocation this should only fail if the the largest free
region is not large enough for the final objects anyways and a GC
needs to happen.

In order to implement this, the allocation hint needs to be passed
down all the way from the factory allocation method to RegionAllocator.

This CL solves this by introducing a V8-internal AllocationHint class
that we pass down from the factory method to the MemoryAllocator.

In the MemoryAllocator we then need to pass the allocation hint
through the V8 API-exposed v8::PageAllocator method. So we can't
use our internal AllocationHint type for this. That's why this CL
also adds a new v8::PageAllocator::AllocationHint class and a new
method v8::PageAllocator::AllocatePages() on the API. This new
method receives the allocation hint as an additional argument.

The BoundedPageAllocator can then use the allocation hint to change
the allocation strategy for RegionAllocator's AllocateRegion().

Bug: 409069381
Change-Id: I60ad4c6da56b2555d86a2f33b7d5f1640329d776
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6495193
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100119}