[fuzzing] Force default for turboshaft flag when fuzzing
Any uses of --no-turboshaft on fuzzers typically runs into a CHECK,
since turboshaft is tied to turbofan. The flag is still picked from
some existing test cases in the test archives. To silence this,
this change ensures the turboshaft flag is reset whenever also
--fuzzing is passed.

If we ever want to make use of --no-turboshaft in isolation again,
this line should be removed again.

Bug: 393401455
Change-Id: I76ef5f3a7138cfae1eb00d974e44371b99472dc2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6520458
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100142}