Add compile flag to make F.p.arguments and .caller accessors
V8 currently implements legacy sloppy mode function.arguments and
function.caller properties as own properties on sloppy, normal
functions. This has the disadvantages of being misaligned with both
Firefox and Safari, and being frequently called because there's a lot of
web code that enumerates own properties of sloppy functions.

This CL adds a compile time flag for exposing these as own properties,
which is false by default. When the flag is false, F.p.arguments and
.caller become ordinary accessors on Function.prototype that throw for
strict or non-normal function receivers. The old behavior can be toggled
by setting the compile time flag to true, if the accessor behavior turns
out to be incompatible.

Bug: 414525205
Change-Id: I7df43cd8b4ea0c8f40363ab206c1f3484580c157
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6502712
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100131}