[test] Fix default kind of code generated by CodeAssemblerTester
The CodeAssemblerTester by default generates code of the BUILTIN kind.
This code, however, is not placed in the read-only space, as should be
the case for builtins. For this reason, in rare circumstances, when
the garbage collector clears non-live references and finds a stale
dispatch handle table entry, a DCHECK fails. This has been witnessed
when conservative stack scanning (CSS) is enabled, because of a false
positive, but it can also happen without CSS, as the introduced test
shows.

This CL fixes the issue by changing the default kind of code generated
by CodeAssemblerTester, from BUILTIN to FOR_TESTING.

Bug: 414753764
Change-Id: Ic2cb8d88f508b636f1238cd2eb5f9a506ca500ee
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6519776
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100145}