[infra] Use prebuilt llvm-symbolizer for arm64
This pulls in a prebuilt symbolizer executable for linux-arm64 from
a build suggested at https://crbug.com/414750472#comment17.

The pinned toolchain currently only supports a symbolizer
executable in the host toolchain (x86).

The executable is reuploaded to V8's storage bucket and downloaded
as a hook only on linux-arm64. GN and test runner are changed to use
this path on linux-arm64.

Bug: 414284086, 414750472
Change-Id: I03cecb870396de355c2b577906f80b39d02996d0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6519361
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100182}