[string] PushStackTraceAndDie with recursion guard
We do want the stack trace for invalid string dispatch, since we want to
know where the bad string came from, but we need to prevent infinite
recursion from printing that stack trace. Add a thread local which
detects the recursion, and dispatches as if we were looking at the
empty string on recursion; this is an invalid dispatch and will likely
result in an OOB read of the empty string, but we're dying anyway and
inside the sandbox this is a safe read.

Bug: 416194305
Change-Id: Idcdbeb9e4a105ac7bf53102e5fe0141533275bff
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6530483
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100184}