[compiler][arm64] Fold shifts into tagged uncompressed ldr and str
In the interpreter, virtual registers are accessed using indices,
scaled to the system pointer size, because we store uncompressed tagged
values on the stack (see `InterpreterAssembler::LoadRegister()`).

However, TurboFan's `MachineType`, as used for loads and stores, does
not differentiate between uncompressed and compressed representations.
They are both "Tagged" and as such are both viewed as 4-bytes wide
accesses, whereas the former is 8-bytes.  Because of this, we miss the
opportunity to fold shifts into those loads and stores.

With TurboShaft, we now have this information, via the
`MemoryRepresentation` type, but it's not used by the instruction
selector when folding shifts, which this CLs changes.

Change-Id: I821d27f38670285c5cb9571651d49fe67f3db786
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6533077
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Pierre Langlois <pierre.langlois@arm.com>
Cr-Commit-Position: refs/heads/main@{#100285}