[wasm][cleanup] Remove dead code related to on-heap wasm-to-js wrappers
In an earlier version of the code, An optimized wasm-to-js wrapper could
be compiled either as an on-heap Code object or as an off-heap WasmCode
object.
We only use the off-heap variant now, so remove some unused code related
to stack frame iteration of the on-heap wrappers. In particular:
- Remove the WASM_TO_JS_FUNCTION stack frame type
- Remove the WasmToJSFunctionFrame class
- Remove the special case in TypedFrame::Iterate()

R=jkummerow@chromium.org
CC=sroettger@chromium.org

Change-Id: I957412f34298b7bea5876aedd0b795c00f142744
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6532352
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100212}