[wasm] Emit write barriers when writing into shared arrays
Shared arrays are allocated in the shared space which needs a write barrier like regular old objects directly after allocation. This CL emits the necessary write barriers for shared arrays in Liftoff.

Note that we could make use of a range barrier in ArrayNew() as a
potential performance improvement since we always write the same
value. If the value doesn't require a write barrier, we could skip
the write barrier checks for the whole array.

Bug: 416437112
Change-Id: I708bddf4dd01e2df3cbad74601fbe195abb3e4e3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6532355
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100253}