[wasm] Use builtins from EmbeddedData for WasmToJs wrappers
With crbug.com/412553113, we want to reuse WasmCodePointers for import
wrappers. Since the code pointers are initialized pointing to the
wrapper builtins and the code pointers can be reused across isolates,
let the entry point to the embedded data instead of the potentially
isolate-owned remapped builtins.

Bug: 412553113
Change-Id: I5521bcbe18792a2a1f3af7e6690504eed4c1f953
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6530022
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Stephen Röttger <sroettger@google.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100234}