[segmented-table] Split segment allocations into smaller segments
Segments are reduced in size and allocated in batches. There is a
segment pool on the table which keeps segments until they are used.
The segment pool is simply an array of atomic offsets which are
consumed using CAS.

Due to the reduced segment size we now need the ability to allocate
multiple read only segments. This is done by pre-reserving the
region up to kNumReadOnlySegments for read only space and extending
it as needed.

To be able to release the segments individually we add a new
AllocatePagesArray on virtual address spaces.

Bug: 411351338
Change-Id: I9e9a4e41b4111ec06114da458701bd21d308f377
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6520233
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100243}