[zone] Harden Zone::AllocateArray
This CL adds an overflow check to ensure that the multiplication of the
requested length with the element size does not overflow. This is a
defense-in-depth measure to avoid potential memory corruption since we
would return a too-small allocation in case of an integer overflow.

Bug: 414831374
Change-Id: I84e3470d7a59b393c6f3631efad9ff181f06c34b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6540007
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100233}