[wasm] track WasmImportWrapperHandles in the WasmImportWrapperCache
Currently, every entry in a dispatch table or WasmJSFunction creates a
separate WasmImportWrapperHandle.

With this change, we now create the WasmImportWrapperHandles inside the
WasmImportWrapperCache. By reusing the WasmCodePointers, tiering up a
single wrapper will tier up all wrappers with the same cache key.

To track the wrappers without keeping them alive, we store a weak_ptr to
the WasmImportWrapperHandle in the cache.

More details in:
https://docs.google.com/document/d/1LDORAyrrGm-xJBw8XfOBWCWKFcoJcmIIq9-op9EGYCc/edit?usp=sharing

Bug: 412553113

Change-Id: Ic66767584bb867c0f5238e96bc4d3d9714a031d7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6495869
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#100258}