[heap] Fix segfault in CSS with scope
The scope was adding 1 to address to the frame address to make sure the
whole frame is scanned. In the fallback to scanning from the stack
start, this causes a segfault since we're reading past the stack. Adding
1 wasn't actually necessary to begin with since the frame pointer
wouldn't point to any interesting V8 objects.
Fix by not adding 1 to the address.

Bug: 379788114
Change-Id: I740f3b3740c99923cf8fb9fd65a74612ec328cc4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6542655
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100261}