Use unsigned integers for HashTable size and capacity
Until now, we used a (signed) int for the NumberOfElements() and
Capacity() of a HashTable. Using signed integers for things like sizes
is generally somewhat dangerous as it can for example lead to seemingly
impossible integer overflows if the (32-bit) int is first promoted to a
(64-bit) size_t, which is then increased further (e.g. through a
multiplication with an element size). This is particularly problematic
for the sandbox as the on-heap integer value is fully
attacker-controlled. To avoid these problems, unsigned integer types
should be used for values that represent object sizes or similar. This
CL therefore switches HashTableBase to use uint32_t instead.

Bug: 414831374
Change-Id: Iba3a8034b124a4452cb843f9b46b41def8dced7f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6540830
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100269}