[serializer] serialize ExternalPointers in InterceptorInfo properly
Previously the ObjectSerializer didn't serialize the ExternalPointers
in the InterceptorInfo properly, but this case can be shadowed by
the fact that they get promoted to RO space by default and don't get
serialized by ObjectSerializer. This patch fixes up the missing
handling in ObjectSerializer and adds a test case for this path.

Refs: https://github.com/nodejs/node/pull/58064

Change-Id: Icc62a01b006eaf68d1d2be1e3bc98b448f0c66dc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6516091
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Joyee Cheung <joyee@igalia.com>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100315}