[turboshaft] Allow GVNing operations that CanRead
Based on an "epoch" (like what Maglev does): GVN has a current_epoch
that is increased on Stores. When GVN records operations that
CanReadMemory, it also records the current_epoch. And when considering
to GVN an operation that CanReadMemory, we only do it if its epoch is
the same as the current_epoch.

Change-Id: Ifeee99bd51e88742dd7e51f36f40788c9b575187
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6531250
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100296}