[wasm-custom-desc] Allow inlining accessors
When inlining the accessor for an accessor property, we materialize
a Call node without type feedback, which hence defaults to disallowing
speculation, which prevents the inlining of Wasm functions in order to
prevent deopt loops caused by unsupported parameter conversions.
This patch tunes the bailout condition to only block inlining if there
is actual feedback that says we shouldn't speculate. This allows
accessors to get inlined, without increasing the risk of deopt loops
for existing use cases.

Bug: 403372470
Change-Id: Ib00593a447812961fba75d9de2f7fdbe94efc123
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6552715
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100300}