[wasm][shared] Allocate WasmArray using kDoubleUnaligned
This removes the artificial alignment padding in the WasmArray header
and instead forces an unaligned allocation of such nature that the body
of the array is aligned. The WasmArray header is not a multiple of 8,
so this requires allocating the object at an address such that
(address % 8 == 4).

On non-pointer-compressed 64 bit platforms this is not the case as
kDoubleUnaligned does not actually adhere to the "un-alignment" but
instead falls back to 8 byte alignment, so there the extra padding in
the header is kept.

Bug: 42204563
Change-Id: I287e090dba47b4b28c6b3d0f2b6af3908841c2cd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6542662
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100322}