[runtime] Reuse JSFunctions in eval
Previously we'd store a native-context specific feedback cell in the
eval cache as a way to share info across eval calls (+ sfi of course).
This CL changes that to storing the entire JSFunction.

When we run the code we'll load the context from the function, and all
accesses to the context go through that. This means that the
JSFunction[[context]] only needs to be correct on the initial call. So
we can reuse the JSFunction and swap the context right before we run
eval.

Functions that are called are tracked in Call feedback though, so we
need to make sure to break consuming that feedback. Otherwise the
optimizing compiler might inline whatever context it sees on the
function it tracked as the canonical context for that function. This
CL updates the heap broker to ignore providing feedback for
script-top-level functions (function literal id 0).

Change-Id: I0d7aeb654fe9b4e03b6a9922b29078261cce99c3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6556163
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100334}